pipeline/requirements.txt
praw>=7.7.0
anthropic>=0.25.0
openai>=1.12.0
httpx>=0.27.0
python-dotenv>=1.0.0
rich>=13.7.0
psycopg2-binary>=2.9.9
pydantic>=2.6.0
